--- src/main/java/wooteco/subway/admin/controller/LineStationsController.java
@@ -1,39 +1,0 @@
-package wooteco.subway.admin.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import wooteco.subway.admin.dto.LineStationCreateRequest;
-import wooteco.subway.admin.service.LineService;
-
-@RestController
-@RequestMapping("/lineStations")
-public class LineStationsController {
-
-    private LineService lineService;
-
-    public LineStationsController(LineService lineService) {
-        this.lineService = lineService;
-    }
-
-    @GetMapping
-    public ResponseEntity showLinesWithStation() {
-        return ResponseEntity.ok().body(lineService.showLinesWithStations());
-    }
-
-    @PostMapping("/{lineId}")
-    public ResponseEntity create(@PathVariable Long lineId, @RequestBody LineStationCreateRequest request) {
-        lineService.addLineStation(lineId, request);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/{lineId}")
-    public ResponseEntity showStationsByLineId(@PathVariable Long lineId) {
-        return ResponseEntity.ok().body(lineService.findLineWithStationsById(lineId));
-    }
-
-    @DeleteMapping("/{lineId}/{stationId}")
-    public ResponseEntity delete(@PathVariable Long lineId, @PathVariable Long stationId) {
-        lineService.removeLineStation(lineId, stationId);
-        return ResponseEntity.noContent().build();
-    }
-}
